requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
transformers
scikit-learn
pyspark
apache-airflow